Changed generation to use Serializer singletons to redeuce template code bloat.

+++ cpp/opendnp3/src/opendnp3/objects/Group11.h
@@ -26,7 +26,7 @@ namespace opendnp3 {
 struct Group11Var1
 {
   static const GroupVariationID ID;
-  static const size_t SIZE = 1;
+  static const uint32_t SIZE = 1;
   static Group11Var1 Read(openpal::ReadOnlyBuffer&);
   static void Write(const Group11Var1&, openpal::WriteBuffer&);
 
@@ -36,7 +36,7 @@ struct Group11Var1
 struct Group11Var2
 {
   static const GroupVariationID ID;
-  static const size_t SIZE = 7;
+  static const uint32_t SIZE = 7;
   static Group11Var2 Read(openpal::ReadOnlyBuffer&);
   static void Write(const Group11Var2&, openpal::WriteBuffer&);
