+++ cpp/opendnp3/src/opendnp3/objects/Group21.h
@@ -26,7 +26,7 @@ namespace opendnp3 {
 struct Group21Var1
 {
   static const GroupVariationID ID;
-  static const size_t SIZE = 5;
+  static const uint32_t SIZE = 5;
   static Group21Var1 Read(openpal::ReadOnlyBuffer&);
   static void Write(const Group21Var1&, openpal::WriteBuffer&);
 
@@ -38,7 +38,7 @@ struct Group21Var1
 struct Group21Var2
 {
   static const GroupVariationID ID;
-  static const size_t SIZE = 3;
+  static const uint32_t SIZE = 3;
   static Group21Var2 Read(openpal::ReadOnlyBuffer&);
   static void Write(const Group21Var2&, openpal::WriteBuffer&);
 
@@ -50,7 +50,7 @@ struct Group21Var2
 struct Group21Var3
 {
   static const GroupVariationID ID;
-  static const size_t SIZE = 5;
+  static const uint32_t SIZE = 5;
   static Group21Var3 Read(openpal::ReadOnlyBuffer&);
   static void Write(const Group21Var3&, openpal::WriteBuffer&);
 
@@ -62,7 +62,7 @@ struct Group21Var3
 struct Group21Var4
 {
   static const GroupVariationID ID;
-  static const size_t SIZE = 3;
+  static const uint32_t SIZE = 3;
   static Group21Var4 Read(openpal::ReadOnlyBuffer&);
   static void Write(const Group21Var4&, openpal::WriteBuffer&);
 
@@ -74,7 +74,7 @@ struct Group21Var4
 struct Group21Var5
 {
   static const GroupVariationID ID;
-  static const size_t SIZE = 11;
+  static const uint32_t SIZE = 11;
   static Group21Var5 Read(openpal::ReadOnlyBuffer&);
   static void Write(const Group21Var5&, openpal::WriteBuffer&);
 
@@ -87,7 +87,7 @@ struct Group21Var5
 struct Group21Var6
 {
   static const GroupVariationID ID;
-  static const size_t SIZE = 9;
+  static const uint32_t SIZE = 9;
   static Group21Var6 Read(openpal::ReadOnlyBuffer&);
   static void Write(const Group21Var6&, openpal::WriteBuffer&);
 
@@ -100,7 +100,7 @@ struct Group21Var6
 struct Group21Var7
 {
   static const GroupVariationID ID;
-  static const size_t SIZE = 11;
+  static const uint32_t SIZE = 11;
   static Group21Var7 Read(openpal::ReadOnlyBuffer&);
   static void Write(const Group21Var7&, openpal::WriteBuffer&);
 
@@ -113,7 +113,7 @@ struct Group21Var7
 struct Group21Var8
 {
   static const GroupVariationID ID;
-  static const size_t SIZE = 9;
+  static const uint32_t SIZE = 9;
   static Group21Var8 Read(openpal::ReadOnlyBuffer&);
   static void Write(const Group21Var8&, openpal::WriteBuffer&);
 
@@ -126,7 +126,7 @@ struct Group21Var8
 struct Group21Var9
 {
   static const GroupVariationID ID;
-  static const size_t SIZE = 2;
+  static const uint32_t SIZE = 2;
   static Group21Var9 Read(openpal::ReadOnlyBuffer&);
   static void Write(const Group21Var9&, openpal::WriteBuffer&);
 
@@ -137,7 +137,7 @@ struct Group21Var9
 struct Group21Var10
 {
   static const GroupVariationID ID;
-  static const size_t SIZE = 4;
+  static const uint32_t SIZE = 4;
   static Group21Var10 Read(openpal::ReadOnlyBuffer&);
   static void Write(const Group21Var10&, openpal::WriteBuffer&);
 
@@ -148,7 +148,7 @@ struct Group21Var10
 struct Group21Var11
 {
   static const GroupVariationID ID;
-  static const size_t SIZE = 2;
+  static const uint32_t SIZE = 2;
   static Group21Var11 Read(openpal::ReadOnlyBuffer&);
   static void Write(const Group21Var11&, openpal::WriteBuffer&);
 
@@ -159,7 +159,7 @@ struct Group21Var11
 struct Group21Var12
 {
   static const GroupVariationID ID;
-  static const size_t SIZE = 4;
+  static const uint32_t SIZE = 4;
   static Group21Var12 Read(openpal::ReadOnlyBuffer&);
   static void Write(const Group21Var12&, openpal::WriteBuffer&);
