+++ cpp/opendnp3/src/opendnp3/objects/Group51.h
@@ -26,7 +26,7 @@ namespace opendnp3 {
 struct Group51Var1
 {
   static const GroupVariationID ID;
-  static const size_t SIZE = 6;
+  static const uint32_t SIZE = 6;
   static Group51Var1 Read(openpal::ReadOnlyBuffer&);
   static void Write(const Group51Var1&, openpal::WriteBuffer&);
 
@@ -36,7 +36,7 @@ struct Group51Var1
 struct Group51Var2
 {
   static const GroupVariationID ID;
-  static const size_t SIZE = 6;
+  static const uint32_t SIZE = 6;
   static Group51Var2 Read(openpal::ReadOnlyBuffer&);
   static void Write(const Group51Var2&, openpal::WriteBuffer&);
