+++ cpp/opendnp3/src/opendnp3/objects/Group52.h
@@ -26,7 +26,7 @@ namespace opendnp3 {
 struct Group52Var1
 {
   static const GroupVariationID ID;
-  static const size_t SIZE = 2;
+  static const uint32_t SIZE = 2;
   static Group52Var1 Read(openpal::ReadOnlyBuffer&);
   static void Write(const Group52Var1&, openpal::WriteBuffer&);
 
@@ -36,7 +36,7 @@ struct Group52Var1
 struct Group52Var2
 {
   static const GroupVariationID ID;
-  static const size_t SIZE = 2;
+  static const uint32_t SIZE = 2;
   static Group52Var2 Read(openpal::ReadOnlyBuffer&);
   static void Write(const Group52Var2&, openpal::WriteBuffer&);
